fixed slight bug in enumerate_range.cpp. cause by difference between 32 and 64 bit operating systems. Fixed by replacing long with ptrdiff_t
git-svn-id: http://origin.googlecode.com/svn/trunk@79 3dbf15f1-5c06-4b2d-4fb8-2ed049db4e7e

+++ range/tests/enumerate_range.cpp
@@ -18,7 +18,7 @@ int main()
 {
   // Note that the type here must agree with the enumerator's value type,
   // otherwise you can get some weird type errors. Note long == ptrdiff_t
-  typedef pair<long, string> Pair;
+  typedef pair<ptrdiff_t, string> Pair;
 
   string str[3] = {"a", "b", "c"};
